feat: 🎸 move the link implementation in commonmark into theme
