+++ packages/preset-commonmark/src/mark/code-inline.ts
@@ -1,5 +1,5 @@
 /* Copyright 2021, Milkdown by Mirone. */
-import { Color, createCmd, createCmdKey, ThemeColor, ThemeFont, ThemeSize } from '@milkdown/core';
+import { createCmd, createCmdKey } from '@milkdown/core';
 import { markRule, toggleMark } from '@milkdown/prose';
 import { createMark, createShortcut } from '@milkdown/utils';
 
@@ -11,20 +11,6 @@ const id = 'code_inline';
 export const ToggleInlineCode = createCmdKey('ToggleInlineCode');
 
 export const codeInline = createMark<Keys>((utils) => {
-    const style = utils.getStyle((themeManager, { css }) => {
-        const palette = (color: Color, opacity = 1) => themeManager.get(ThemeColor, [color, opacity]);
-        const radius = themeManager.get(ThemeSize, 'radius');
-        const code = themeManager.get(ThemeFont, 'code');
-
-        return css`
-            background-color: ${palette('neutral')};
-            color: ${palette('background')};
-            border-radius: ${radius};
-            font-weight: 500;
-            font-family: ${code};
-            padding: 0 0.2rem;
-        `;
-    });
     return {
         id,
         schema: () => ({
@@ -32,7 +18,7 @@ export const codeInline = createMark<Keys>((utils) => {
             code: true,
             inclusive: false,
             parseDOM: [{ tag: 'code' }],
-            toDOM: (mark) => ['code', { class: utils.getClassName(mark.attrs, 'code-inline', style) }],
+            toDOM: (mark) => ['code', { class: utils.getClassName(mark.attrs, 'code-inline') }],
             parseMarkdown: {
                 match: (node) => node.type === 'inlineCode',
                 runner: (state, node, markType) => {
