+++ packages/theme-nord/src/icon.ts
